refactor : refactor the code
