+++ .env
@@ -1,1 +1,1 @@
-REACT_APP_JWT_SECRET=<your-jwt-secret>
+REACT_APP_JWT_SECRET=ecommerceapp
